introduce updating status

+++ app/Actions/HostedDomain/DeleteHostedDomain.php
@@ -16,6 +16,12 @@ class DeleteHostedDomain
             ]);
         }
 
+        if ($hostedDomain->status->isProcessing()) {
+            throw ValidationException::withMessages([
+                'domain' => ['Cannot delete a domain while it is '.$hostedDomain->status->value.'.'],
+            ]);
+        }
+
         $hostedDomain->delete();
     }
 }
